Hide community side bar from Help Center
Enabling the community even with only internally visible topics caused
the title and column to show up on the main page and search results
which could potentially confuse Customers and will be revisited later
on should we expose the Community threads to logged in users.  Also
fixed some iconography that was missing in the default template.